# Conflicts:
#	GeneralsMD/Code/Main/WinMain.cpp

+++ Generals/Code/GameEngine/Source/GameLogic/Object/Update/ParticleUplinkCannonUpdate.cpp
@@ -56,7 +56,6 @@
 #include "GameLogic/Module/SpecialPowerModule.h"
 #include "GameLogic/Module/ParticleUplinkCannonUpdate.h"
 #include "GameLogic/Module/PhysicsUpdate.h"
-#include "GameLogic/Module/LaserUpdate.h"
 #include "GameLogic/Module/ActiveBody.h"
 
 #ifdef RTS_INTERNAL
@@ -212,7 +211,7 @@ void ParticleUplinkCannonUpdate::killEverything()
 	removeAllEffects();
 
 	//This laser is independent from the other effects and needs to be specially handled.
-	if( m_orbitToTargetBeamID )
+	if( m_orbitToTargetBeamID != INVALID_DRAWABLE_ID )
 	{
 		Drawable *beam = TheGameClient->findDrawableByID( m_orbitToTargetBeamID );
 		if( beam )
@@ -221,6 +220,7 @@ void ParticleUplinkCannonUpdate::killEverything()
 		}
 		m_orbitToTargetBeamID = INVALID_DRAWABLE_ID;
 	}
+	m_orbitToTargetLaserRadius = LaserRadiusUpdate();
 
 	TheAudio->removeAudioEvent( m_powerupSound.getPlayingHandle() );
 	TheAudio->removeAudioEvent( m_unpackToReadySound.getPlayingHandle() );
@@ -410,38 +410,43 @@ UpdateSleepTime ParticleUplinkCannonUpdate::update()
 				break;
 			case LASERSTATUS_BORN:
 			{
-				Drawable *beam = TheGameClient->findDrawableByID( m_orbitToTargetBeamID );
-				if( beam )
+				if( orbitalDecayStart <= now )
 				{
-					//m_annihilationSound.setPosition( beam->getPosition() );
-					if( orbitalDecayStart <= now )
+					Drawable *beam = TheGameClient->findDrawableByID( m_orbitToTargetBeamID );
+					if( beam )
 					{
+						//m_annihilationSound.setPosition( beam->getPosition() );
 						static NameKeyType nameKeyClientUpdate = NAMEKEY( "LaserUpdate" );
 						LaserUpdate *update = (LaserUpdate*)beam->findClientUpdateModule( nameKeyClientUpdate );
 						if( update )
 						{
 							update->setDecayFrames( data->m_widthGrowFrames );
 						}
-						m_laserStatus = LASERSTATUS_DECAYING;
 					}
+					m_orbitToTargetLaserRadius.setDecayFrames( data->m_widthGrowFrames );
+					m_laserStatus = LASERSTATUS_DECAYING;
 				}
 				break;
 			}
 			case LASERSTATUS_DECAYING:
 			{
-				Drawable *beam = TheGameClient->findDrawableByID( m_orbitToTargetBeamID );
-				if( beam )
+				if( orbitalDeathFrame <= now )
 				{
-					//m_annihilationSound.setPosition( beam->getPosition() );
 					TheAudio->removeAudioEvent( m_annihilationSound.getPlayingHandle() );
-					if( orbitalDeathFrame <= now )
+					if ( m_orbitToTargetBeamID != INVALID_DRAWABLE_ID )
 					{
-						TheGameClient->destroyDrawable( beam );
+						Drawable *beam = TheGameClient->findDrawableByID( m_orbitToTargetBeamID );
+						if( beam )
+						{
+							//m_annihilationSound.setPosition( beam->getPosition() );
+							TheGameClient->destroyDrawable( beam );
+						}
 						m_orbitToTargetBeamID = INVALID_DRAWABLE_ID;
-						m_laserStatus = LASERSTATUS_DEAD;
-						m_startAttackFrame = 0;
-						setLogicalStatus( STATUS_IDLE );
 					}
+					m_orbitToTargetLaserRadius = LaserRadiusUpdate();
+					m_laserStatus = LASERSTATUS_DEAD;
+					m_startAttackFrame = 0;
+					setLogicalStatus( STATUS_IDLE );
 				}
 				break;
 			}
@@ -450,8 +455,7 @@ UpdateSleepTime ParticleUplinkCannonUpdate::update()
 				break;
 		}
 
-		Drawable *beam = TheGameClient->findDrawableByID( m_orbitToTargetBeamID );
-		if( beam && orbitalBirthFrame <= now && now <= orbitalDeathFrame )
+		if( orbitalBirthFrame <= now && now < orbitalDeathFrame )
 		{
 		
 			if( !m_manualTargetMode )
@@ -546,16 +550,35 @@ UpdateSleepTime ParticleUplinkCannonUpdate::update()
 
 			Real scorchRadius = 0.0f;
 			Real damageRadius = 0.0f;
+			Real templateLaserRadius = 13.0f;
+			Real visualLaserRadius = 0.0f;
 
 			//Reset the laser position
-			static NameKeyType nameKeyClientUpdate = NAMEKEY( "LaserUpdate" );
-			LaserUpdate *update = (LaserUpdate*)beam->findClientUpdateModule( nameKeyClientUpdate );
-			if( update )
+			Drawable *beam = TheGameClient->findDrawableByID( m_orbitToTargetBeamID );
+			if ( beam )
 			{
-				update->initLaser( NULL, &orbitPosition, &m_currentTargetPosition );
-				scorchRadius = update->getCurrentLaserRadius() * data->m_scorchMarkScalar;
-				damageRadius = update->getCurrentLaserRadius() * data->m_damageRadiusScalar;
+				static NameKeyType nameKeyClientUpdate = NAMEKEY( "LaserUpdate" );
+				LaserUpdate *update = (LaserUpdate*)beam->findClientUpdateModule( nameKeyClientUpdate );
+				if( update )
+				{
+					update->initLaser( NULL, &orbitPosition, &m_currentTargetPosition );
+					// TheSuperHackers @logic-client-separation The GameLogic has a dependency on this drawable.
+					// The logical laser radius for the damage should probably be part of ParticleUplinkCannonUpdateModuleData.
+					templateLaserRadius = update->getTemplateLaserRadius();
+					visualLaserRadius = update->getCurrentLaserRadius();
+				}
 			}
+			// TheSuperHackers @refactor helmutbuhler/xezon 17/05/2025
+			// Originally the damageRadius was calculated with a value updated by LaserUpdate::clientUpdate.
+			// To no longer rely on GameClient updates, this class now maintains a copy of the LaserRadiusUpdate.
+			m_orbitToTargetLaserRadius.updateRadius();
+			const Real logicalLaserRadius = templateLaserRadius * m_orbitToTargetLaserRadius.getWidthScale();
+			damageRadius = logicalLaserRadius * data->m_damageRadiusScalar;
+			scorchRadius = logicalLaserRadius * data->m_scorchMarkScalar;
+#if defined(RETAIL_COMPATIBLE_CRC)
+			DEBUG_ASSERTCRASH(logicalLaserRadius == visualLaserRadius,
+				("ParticleUplinkCannonUpdate's laser radius does not match LaserUpdate's laser radius - will cause mismatch in VS6 retail compatible builds\n"));
+#endif
 
 			//Create scorch marks periodically
 			if( m_nextScorchMarkFrame <= now )
@@ -904,11 +927,14 @@ void ParticleUplinkCannonUpdate::createOrbitToTargetLaser( UnsignedInt growthFra
 {
 	const ParticleUplinkCannonUpdateModuleData *data = getParticleUplinkCannonUpdateModuleData();
 
-	Drawable *beam = TheGameClient->findDrawableByID( m_orbitToTargetBeamID );
-	if( beam )
+	if ( m_orbitToTargetBeamID != INVALID_DRAWABLE_ID )
 	{
-		TheAudio->removeAudioEvent( m_annihilationSound.getPlayingHandle() );
-		TheGameClient->destroyDrawable( beam );
+		Drawable *beam = TheGameClient->findDrawableByID( m_orbitToTargetBeamID );
+		if( beam )
+		{
+			TheAudio->removeAudioEvent( m_annihilationSound.getPlayingHandle() );
+			TheGameClient->destroyDrawable( beam );
+		}
 		m_orbitToTargetBeamID = INVALID_DRAWABLE_ID;
 	}
 
@@ -939,6 +965,9 @@ void ParticleUplinkCannonUpdate::createOrbitToTargetLaser( UnsignedInt growthFra
 			m_annihilationSound.setPlayingHandle( TheAudio->addAudioEvent( &m_annihilationSound ) );
 		}
 	}
+
+	m_orbitToTargetLaserRadius = LaserRadiusUpdate();
+	m_orbitToTargetLaserRadius.initRadius( growthFrames );
 }
 
 //-------------------------------------------------------------------------------------------------
@@ -1282,7 +1311,11 @@ void ParticleUplinkCannonUpdate::xfer( Xfer *xfer )
 	const ParticleUplinkCannonUpdateModuleData *data = getParticleUplinkCannonUpdateModuleData();
 
 	// version
-	XferVersion currentVersion = 2;
+#if RETAIL_COMPATIBLE_XFER_SAVE
+	const XferVersion currentVersion = 2;
+#else
+	const XferVersion currentVersion = 4;
+#endif
 	XferVersion version = currentVersion;
 	xfer->xferVersion( &version, currentVersion );
 
@@ -1383,6 +1416,11 @@ void ParticleUplinkCannonUpdate::xfer( Xfer *xfer )
 	// the time of the 2nd last manual target click
 	xfer->xferUnsignedInt( &m_2ndLastDrivingClickFrame );
 
+	if( version >= 4 )
+	{
+		m_orbitToTargetLaserRadius.xfer( xfer );
+	}
+
 }  // end xfer
 
 // ------------------------------------------------------------------------------------------------
@@ -1393,5 +1431,28 @@ void ParticleUplinkCannonUpdate::loadPostProcess( void )
 
 	// extend base class
 	UpdateModule::loadPostProcess();
+
+#if RETAIL_COMPATIBLE_XFER_SAVE
+	// TheSuperHackers @info xezon 17/05/2025
+	// For retail game compatibility, this transfers the loaded visual laser radius
+	// settings from the Drawable's LaserUpdate to the local LaserRadiusUpdate.
+	if( m_orbitToTargetBeamID != INVALID_DRAWABLE_ID )
+	{
+		Drawable* drawable = TheGameClient->findDrawableByID( m_orbitToTargetBeamID );
+		if( drawable != NULL )
+		{
+			static NameKeyType nameKeyClientUpdate = NAMEKEY( "LaserUpdate" );
+			LaserUpdate *update = (LaserUpdate*)drawable->findClientUpdateModule( nameKeyClientUpdate );
+			if( update != NULL )
+			{
+				m_orbitToTargetLaserRadius = update->getLaserRadiusUpdate();
+			}
+		}
+		else
+		{
+			DEBUG_CRASH(( "ParticleUplinkCannonUpdate::loadPostProcess - Unable to find drawable for m_orbitToTargetBeamID\n" ));
+		}
+	}
+#endif
 
 }  // end loadPostProcess
